\name{mtkBasicMonteCarloDesignerResult}
\alias{mtkBasicMonteCarloDesignerResult}
\title{The constructor of class \code{mtkBasicMonteCarloDesignerResult}}

\description{The constructor}

\usage{mtkBasicMonteCarloDesignerResult(main,information=NULL)
}
\value{an object of the \code{\linkS4class{mtkBasicMonteCarloDesignerResult}} class}

\arguments{
\item{main}{a data.frame holding the experimental design produced by the designer.}
\item{information}{a named list  containing the information about the managed data and the underlying process.}
}

\author{Juhui WANG, MIA-Jouy, Inra,  Juhui.Wang@jouy.inra.fr}
\references{
\enumerate{
\item A. Saltelli, K. Chan and E. M. Scott (2000). Sensitivity Analysis. Wiley, New York.
\item J. Wang, H. Richard, R. Faivre, H. Monod (2013). Le package \code{mtk}, une bibliothèque R pour l'exploration numérique des modèles.
 \emph{In:} Analyse de sensibilité et exploration de modèles :  Application aux sciences de la nature et  de l'environnement 
 (R. Faivre, B. Iooss, S. Mahévas, D. Makowski, H. Monod, Eds). Editions Quae, Versailles.
}
}

\examples{

# see examples with help(BasicMonteCarlo)

}

