\name{mtkFastDesignerResult}
\alias{mtkFastDesignerResult}
\title{The constructor of the class \code{mtkFastDesignerResult}}

\description{The constructor}

\usage{mtkFastDesignerResult(main,information=NULL)
}
\value{an object of the \code{\linkS4class{mtkFastDesignerResult}} class}

\arguments{
\item{main}{a data.frame holding the experimental design produced by the designer.}
\item{information}{a named list  containing the information about the managed data.}
}
\author{Juhui WANG, MIA-Jouy, Inra,  Juhui.Wang@jouy.inra.fr}
\references{J. Wang, H. Richard, R. Faivre, H. Monod (2013). Le package \code{mtk}, une bibliothèque R pour l'exploration numérique des modèles.
 \emph{In:} Analyse de sensibilité et exploration de modèles :  Application aux sciences de la nature et  de l'environnement 
 (R. Faivre, B. Iooss, S. Mahévas, D. Makowski, H. Monod, Eds). Editions Quae, Versailles.}

\examples{

# See examples from the help of the method: help(Fast)
}

