\name{mtkRandLHSDesigner}
\alias{mtkRandLHSDesigner}
\title{The constructor of the class \code{mtkRandLHSDesigner}}


\description{The constructor}
\usage{mtkRandLHSDesigner(mtkParameters = NULL, listParameters = NULL)
}
\value{an object of the \code{\linkS4class{mtkRandLHSDesigner}} class}

\arguments{
\item{mtkParameters}{a vector of [\code{\linkS4class{mtkParameter}}] representing the parameters necessary to run the process.}
\item{listParameters}{a named list  containing the parameters to pass while calling the process. This gives another way to specify the parameters.}
}

\seealso{\code{package?lsh}, \code{help(LHS)}}

\examples{

# To do, example for LHS method
}

