\name{mtkRandLHSDesignerResult}
\alias{mtkRandLHSDesignerResult}
\title{The constructor of the class \code{mtkRandLHSDesignerResult}}

\description{The constructor}

\usage{mtkRandLHSDesignerResult(main,information=NULL)
}
\value{an object of the \code{\linkS4class{mtkRandLHSDesignerResult}} class}

\arguments{
\item{main}{a data.frame holding the experimental design produced by the designer.}
\item{information}{a named list  containing the information about the managed data.}
}
\seealso{\code{package?lsh}, \code{help(LHS)}}

\examples{

# To do, example for LHS method
}
