\name{mtkReadFactors-methods}
\alias{mtkReadFactors-methods}
\alias{mtkReadFactors}
\title{The \code{mtkReadFactor} method}
\description{a list of factors}
\usage{mtkReadFactors(file, path)}
\value{an object of the class \code{mtkDomain}}
\arguments{
\item{file}{the name of the file to read.}
\item{path}{the path to the file to read.}
}
\author{Hervé Richard, BioSP, INRA, Domaine Saint paul, 84914 Avignon Cedex 9}
\examples{
# see examples for the \code{\linkS4class{mtkExpFactors}} class.
}
