\name{mtkRegressionAnalyserResult}
\alias{mtkRegressionAnalyserResult}
\title{The constructor of the class \code{mtkRegressionAnalyserResult}}

\description{The constructor}

\usage{mtkRegressionAnalyserResult(main,information=NULL)
}
\value{an object of the \code{\linkS4class{mtkRegressionAnalyserResult}} class}

\arguments{
\item{main}{a data.frame holding the results of the sensitivity analysis produced by the analyser.}
\item{information}{a named list  containing the information about the managed data.}
}

\seealso{\code{help(morris, sensitivity)} and \code{help(Regression)}}

\examples{

## See examples from help(mtkAnalyserResult)

	
}

