\name{mtkResult}
\alias{mtkResult}
\title{The constructor of the class \code{mtkResult}}

\description{The constructor}

\usage{mtkResult(information=list())
}
\value{an object of the \code{\linkS4class{mtkResult}} class}

\arguments{
\item{information}{a named list  containing the information about the managed data.}
}
\section{Details}{
The \code{\linkS4class{mtkResult}} class is a virtual class to manage the generic properties of results produced by the processes involved in the "mtk" package.

For details uses, see examples from \code{help(mtkAnalyserResult)}, \code{help(mtkDesignerResult)}, \code{help(mtkEvaluatorResult)}.
}
\seealso{\code{help(mtkAnalyserResult)}, \code{help(mtkDesignerResult)}, \code{help(mtkEvaluatorResult)}}
\author{Juhui WANG, MIA-Jouy, Inra, Juhui.Wang@jouy.inra.fr }

\references{J. Wang, H. Richard, R. Faivre, H. Monod (2013). Le package \code{mtk}, une bibliothèque R pour l'exploration numérique des modèles.
 \emph{In:} Analyse de sensibilité et exploration de modèles :  Application aux sciences de la nature et  de l'environnement 
 (R. Faivre, B. Iooss, S. Mahévas, D. Makowski, H. Monod, Eds). Editions Quae, Versailles.}

\examples{

## See examples from help(mtkAnalyserResult), help(mtkDesignerResult), help(mtkEvaluatorResult)

	
}
