\name{mtkSobolAnalyserResult}
\alias{mtkSobolAnalyserResult}
\title{The constructor of the class \code{mtkSobolAnalyserResult}}

\description{The constructor}

\usage{mtkSobolAnalyserResult(main,information=NULL)
}
\value{an object of the \code{\linkS4class{mtkSobolAnalyserResult}} class}

\arguments{
\item{main}{a data.frame holding the results of the sensitivity analysis produced by the analyser.}
\item{information}{a named list  containing the information about the managed data.}
}

\seealso{\code{help(mtkAnalyserResult)} and \code{help(Sobol)}}

\examples{

## See examples from help(mtkAnalyserResult).
	
}
