\name{mtkSobolDesignerResult}
\alias{mtkSobolDesignerResult}
\title{The constructor of the class \code{mtkSobolDesignerResult}}

\description{The constructor}

\usage{mtkSobolDesignerResult(main,information=NULL)
}
\value{an object of the \code{\linkS4class{mtkSobolDesignerResult}} class}

\arguments{
\item{main}{a data.frame holding the experimental design produced by the designer.}
\item{information}{a named list  containing the information about the managed data.}
}

\seealso{\code{help(mtkDesignerResult)} and \code{help(Sobol)}}

\examples{

## See examples from help(mtkDesignerResult).
	
}
