\name{mtkSystemEvaluatorResult}
\alias{mtkSystemEvaluatorResult}
\title{The constructor of the class \code{mtkSystemEvaluatorResult}}

\description{The constructor}

\usage{mtkSystemEvaluatorResult(main,information=NULL)
}
\value{an object of the \code{\linkS4class{mtkSystemEvaluatorResult}} class}

\arguments{
\item{main}{a data.frame holding the results produced by the evaluator.}
\item{information}{a named list  containing the information about the managed data.}
}

\author{Juhui WANG, MIA-Jouy, Inra, Juhui.Wang@jouy.inra.fr }

\examples{
# See examples 
}
