\name{mtkWWDMEvaluatorResult}
\alias{mtkWWDMEvaluatorResult}
\title{The constructor of the class \code{mtkWWDMEvaluatorResult}}

\description{The constructor}

\usage{mtkWWDMEvaluatorResult(main,information=NULL)
}
\value{an object of the \code{\linkS4class{mtkWWDMEvaluatorResult}} class}

\arguments{
\item{main}{a data.frame holding the results produced by the evaluator.}
\item{information}{a named list  containing the information about the managed data.}
}
\seealso{\code{help(mtkEvaluatorResult)} and \code{help(WWDM)}}

\examples{

## See examples from help(mtkEvaluatorResult).
	
}
