\name{serializeOn-methods}
\alias{serializeOn-methods}
\alias{serializeOn}
\title{The \code{serializeOn} method}
\description{Returns all data and informations managed by an object as a named list.}
\usage{serializeOn(this)}
\value{a named list}
\arguments{
\item{this}{the underlying object}
}
\author{Juhui WANG, MIA-Jouy, Inra, Juhui.Wang@jouy.inra.fr }

\examples{
# Function not used yet in the current release.
}
