\name{setReady-methods}
\alias{setReady-methods}
\alias{setReady}
\title{The \code{setReady} method}
\description{Makes the process ready to run.}
\usage{setReady(this, switch)}
\value{invisble()}
\arguments{
\item{this}{the underlying object of the class \code{\linkS4class{mtkProcess}}}
\item{switch}{a logical (TRUE or FALSE).}
}
\section{Details}{
This function is only useful for the programmers who need to program the mtk's internal functions.
}
\author{Juhui WANG, MIA-Jouy, Inra, Juhui.Wang@jouy.inra.fr }

\references{J. Wang, H. Richard, R. Faivre, H. Monod (2013). Le package \code{mtk}, une bibliothèque R pour l'exploration numérique des modèles.
 \emph{In:} Analyse de sensibilité et exploration de modèles :  Application aux sciences de la nature et  de l'environnement 
 (R. Faivre, B. Iooss, S. Mahévas, D. Makowski, H. Monod, Eds). Editions Quae, Versailles.}

\examples{

# This function is only useful for the programmers
#  who need to program the mtk's internal functions.

}


